+++ Rhinoceros_Engine/Create/Point3dList.cs
@@ -32,33 +32,12 @@ namespace BH.Engine.Rhinoceros
         /**** Public Methods                            ****/
         /***************************************************/
 
-        public static List<Curve> RandomCurves(Random random)
-        {
-            return new List<Curve>()
-            {
-                RandomArcCurve(random),
-                RandomLineCurve(random),
-                RandomNurbsCurve(random),
-                RandomPolylineCurve(random),
-                RandomPolyCurve(random),
-            };
-        }
-
-        /***************************************************/
-
         public static List<Point3d> RandomPoint3dList(Random random)
         {
             List<Point3d> points = new List<Point3d>();
             for (int i = 0; i < 10; i++)
                 points.Add(RandomPoint3d(random));
             return points;
-        }
-
-        /***************************************************/
-
-        public static List<Curve> RandomCurves(int  seed = 0)
-        {
-            return RandomCurves(new Random(seed));
         }
 
         /***************************************************/
