+++ Rhinoceros_Engine/Rhinoceros_Engine.csproj
@@ -80,8 +80,10 @@
     <Compile Include="Convert\ToBHoM.cs" />
     <Compile Include="Convert\ToRhino.cs" />
     <Compile Include="Create\Arc.cs" />
+    <Compile Include="Create\ArcCurve.cs" />
     <Compile Include="Create\Circle.cs" />
-    <Compile Include="Create\Collections.cs" />
+    <Compile Include="Create\Point3dList.cs" />
+    <Compile Include="Create\Curves.cs" />
     <Compile Include="Create\Geometry.cs" />
     <Compile Include="Create\Line.cs" />
     <Compile Include="Create\NurbsCurve.cs" />
